+++ test/test_annotations.py
@@ -55,7 +55,7 @@ def test_goto_action(tmp_path):
     x, y, text = 80, 140, "GoTo action"
     pdf.text(x=x, y=y, txt=text)
     pdf.add_action(
-        GoToAction(dest=DestinationXYZ(page=2).dest_str(pdf)),
+        GoToAction(dest=DestinationXYZ(page=2).as_str(pdf)),
         x=x,
         y=y - pdf.font_size,
         w=pdf.get_string_width(text),
@@ -72,7 +72,7 @@ def test_goto_remote_action(tmp_path):
     pdf.add_page()
     x, y, text = 80, 140, "GoTo-Remote action"
     pdf.text(x=x, y=y, txt=text)
-    dest = DestinationXYZ(page=1, page_as_obj_id=False).dest_str(pdf)
+    dest = DestinationXYZ(page=1, page_as_obj_id=False).as_str(pdf)
     pdf.add_action(
         GoToRemoteAction("goto_action.pdf", dest=dest),
         x=x,